✨ Feat: 인증/인가 추가

+++ src/main/kotlin/moomoo/todo/domain/todos/service/TodoServiceImpl.kt
@@ -5,7 +5,7 @@ import moomoo.todo.domain.comments.model.Comment
 import moomoo.todo.domain.comments.model.toResponse
 import moomoo.todo.domain.comments.repository.CommentRepository
 import moomoo.todo.domain.exception.ModelNotFoundException
-import moomoo.todo.domain.exception.PasswordNotMatchedException
+import moomoo.todo.domain.exception.UnauthorizedAccessException
 import moomoo.todo.domain.todos.dto.CreateTodoRequest
 import moomoo.todo.domain.todos.dto.TodoResponse
 import moomoo.todo.domain.todos.dto.TodoResponseWithCommentList
@@ -14,20 +14,26 @@ import moomoo.todo.domain.todos.model.Todo
 import moomoo.todo.domain.todos.model.toResponse
 import moomoo.todo.domain.todos.model.toResponseWithCommentList
 import moomoo.todo.domain.todos.repository.TodoRepository
+import moomoo.todo.domain.users.repository.UserRepository
+import moomoo.todo.infra.security.UserPrincipal
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class TodoServiceImpl(
     private val todoRepository: TodoRepository,
-    private val commentRepository: CommentRepository
-): TodoService {
+    private val commentRepository: CommentRepository,
+    private val userRepository: UserRepository
+) : TodoService {
 
+    @Transactional
     override fun getAllTodoList(): List<TodoResponse> {
         return todoRepository.findAllByOrderByCreatedAtDesc().map { it.toResponse() }
     }
 
+    @Transactional
     override fun getTodoById(todoId: Long): TodoResponseWithCommentList {
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
 
@@ -38,10 +44,10 @@ class TodoServiceImpl(
 
     @Transactional
     override fun createTodo(request: CreateTodoRequest): TodoResponse {
+        val user = userRepository.findByIdOrNull(request.userId) ?: throw ModelNotFoundException("User", request.userId)
+
         val todo = Todo(
-            title = request.title,
-            name = request.name,
-            description = request.description
+            title = request.title, user = user, description = request.description
         )
 
         return todoRepository.save(todo).toResponse()
@@ -51,7 +57,11 @@ class TodoServiceImpl(
     override fun updateTodo(todoId: Long, request: UpdateTodoRequest): TodoResponse {
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
 
-        todo.updateTodo(request.title, request.name, request.description)
+        val userPrincipal = SecurityContextHolder.getContext().authentication.principal as UserPrincipal
+
+        if (userPrincipal.id != todo.user.id) throw UnauthorizedAccessException("todo")
+
+        todo.updateTodo(request.title, request.description)
 
         return todo.toResponse()
     }
@@ -59,6 +69,10 @@ class TodoServiceImpl(
     @Transactional
     override fun deleteTodo(todoId: Long) {
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+
+        val userPrincipal = SecurityContextHolder.getContext().authentication.principal as UserPrincipal
+
+        if (userPrincipal.id != todo.user.id) throw UnauthorizedAccessException("todo")
 
         commentRepository.deleteAllByTodoId(todoId)
 
@@ -69,21 +83,28 @@ class TodoServiceImpl(
     override fun toggleTodo(todoId: Long): TodoResponse {
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
 
+        val userPrincipal = SecurityContextHolder.getContext().authentication.principal as UserPrincipal
+
+        if (userPrincipal.id != todo.user.id) throw UnauthorizedAccessException("todo")
+
         todo.toggleTodo()
 
         return todo.toResponse()
     }
 
+    @Transactional
     override fun getCommentList(todoId: Long): List<CommentResponse> {
-        if(!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
+        if (!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
 
         return commentRepository.findAllByTodoIdOrderByCreatedAt(todoId).map { it.toResponse() }
     }
 
+    @Transactional
     override fun getComment(todoId: Long, commentId: Long): CommentResponse {
-        if(!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
+        if (!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
 
-        val comment = commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        val comment =
+            commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
         return comment.toResponse()
     }
@@ -91,12 +112,10 @@ class TodoServiceImpl(
     @Transactional
     override fun addComment(todoId: Long, request: CreateCommentRequest): CommentResponse {
         val todo = todoRepository.findByIdOrNull(todoId) ?: throw ModelNotFoundException("Todo", todoId)
+        val user = userRepository.findByIdOrNull(request.userId) ?: throw ModelNotFoundException("User", request.userId)
 
         val comment = Comment(
-            name = request.name,
-            comment = request.comment,
-            password = request.password,
-            todo = todo
+            comment = request.comment, user = user, todo = todo
         )
 
         return commentRepository.save(comment).toResponse()
@@ -104,24 +123,30 @@ class TodoServiceImpl(
 
     @Transactional
     override fun updateComment(todoId: Long, commentId: Long, request: UpdateCommentRequest): CommentResponse {
-        if(!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
+        if (!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
 
-        val comment = commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        val comment =
+            commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        if (!comment.isValidPassword(request.password)) throw PasswordNotMatchedException()
+        val userPrincipal = SecurityContextHolder.getContext().authentication.principal as UserPrincipal
 
-        comment.updateComment(request.name, request.comment)
+        if (userPrincipal.id != comment.user.id) throw UnauthorizedAccessException("comment")
+
+        comment.updateComment(request.comment)
 
         return comment.toResponse()
     }
 
     @Transactional
-    override fun deleteComment(todoId: Long, commentId: Long, request: DeleteCommentRequest) {
-        if(!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
+    override fun deleteComment(todoId: Long, commentId: Long) {
+        if (!todoRepository.existsById(todoId)) throw ModelNotFoundException("Todo", todoId)
 
-        val comment = commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        val comment =
+            commentRepository.findByTodoIdAndId(todoId, commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        if (!comment.isValidPassword(request.password)) throw PasswordNotMatchedException()
+        val userPrincipal = SecurityContextHolder.getContext().authentication.principal as UserPrincipal
+
+        if (userPrincipal.id != comment.user.id) throw UnauthorizedAccessException("comment")
 
         commentRepository.delete(comment)
     }
